+++ next.config.ts
@@ -2,9 +2,6 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   distDir: process.env.NODE_ENV === "development" ? ".next-dev" : ".next",
-  experimental: {
-    typedRoutes: true,
-  },
 };
 
 export default nextConfig;